+++ tower/CHANGELOG.md
@@ -9,10 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- **balance**: Remove redundant `Clone` bounds from `Clone` impls
-  for `MakeBalance`, and `MakeBalanceLayer` ([#606])
+- **balance**: Remove redundant `Req: Clone` bound from `Clone` impls
+  for `MakeBalance`, and `MakeBalanceLayer` ([#607])
+- **balance**: Remove redundant `Req: Debug` bound from `Debug` impls
+  for `MakeBalance`, `MakeFuture`, `Balance`, and `Pool` ([#607])
+- **ready-cache**: Remove redundant `Req: Debug` bound from `Debug` impl
+  for `ReadyCache` ([#607])
+- **steer**: Remove redundant `Req: Debug` bound from `Debug` impl
+  for `Steer` ([#607])
 
-[#606]: https://github.com/tower-rs/tower/pull/606
+[#607]: https://github.com/tower-rs/tower/pull/607
 
 # 0.4.10 (October 19, 2021)